+++ Encoding/x265Enc.vb
@@ -949,7 +949,9 @@ Public Class x265Params
                     New NumParam With {.Switch = "--seek", .Text = "Seek"},
                     New BoolParam With {.Switch = "--dither", .Text = "Dither (High Quality Downscaling)"},
                     New BoolParam With {.Switch = "--field", .NoSwitch = "--no-field", .Text = "Field Coding"})
-                Add("Loop Filter", Deblock, DeblockA, DeblockB, SAO,
+                Add("Loop Filter", Deblock, DeblockA, DeblockB,
+                    New NumParam With {.Switch = "--selective-sao", .Text = "Selective SAO", .Init = 4, .Config = {0, 4}},
+                    SAO,
                     New BoolParam With {.Switch = "--limit-sao", .Text = "Limit Sample Adaptive Offset"},
                     SAOnonDeblock)
                 Add("Other",